+++ ecs/query.go
@@ -102,6 +102,22 @@ func (q *Query) Entity() Entity {
 }
 
 // EntityAt returns the entity at a given index.
+//
+// Do not use this to iterate a query! Use [Query.Next] instead.
+//
+// The method is particularly useful for random sampling of entities
+// from a query (see the example).
+// However, performance heavily depends on the number of archetypes in the world and in the query.
+// It is recommended to use a filter which is registered via the [World.Cache],
+// This increases performance at least by factor 4.
+//
+// Some numbers for comparison:
+//   - 1 archetype, filter not registered: ≈ 11ns
+//   - 1 archetype, filter registered: ≈ 3ns
+//   - 10 archetypes, filter not registered: ≈ 50ns
+//   - 10 archetypes, filter registered: ≈ 10ns
+//
+// Panics if the index is out of range, as indicated by [Query.Count].
 func (q *Query) EntityAt(index int) Entity {
 	return q.entityAt(index)
 }
@@ -334,21 +350,21 @@ func (q *Query) stepArchetype(step uint32) (int, bool) {
 func (q *Query) countEntities() int {
 	var count uint32 = 0
 
+	if q.isFiltered {
+		ln := int32(len(q.archetypes))
+		var i int32
+		for i = 0; i < ln; i++ {
+			count += q.archetypes[i].Len()
+		}
+		return int(count)
+	}
+
 	if q.isBatch {
 		batch := q.nodeArchetypes.(*batchArchetypes)
 		nArch := batch.Len()
 		var j int32
 		for j = 0; j < nArch; j++ {
 			count += batch.EndIndex[j] - batch.StartIndex[j]
-		}
-		return int(count)
-	}
-
-	if q.isFiltered {
-		ln := int32(len(q.archetypes))
-		var i int32
-		for i = 0; i < ln; i++ {
-			count += q.archetypes[i].Len()
 		}
 		return int(count)
 	}
@@ -361,8 +377,7 @@ func (q *Query) countEntities() int {
 		if !nd.HasRelation {
 			// There should be at least one archetype.
 			// Otherwise, the node would be inactive.
-			arches := nd.Archetypes()
-			arch := arches.Get(0)
+			arch := nd.Archetypes().Get(0)
 			count += arch.Len()
 			continue
 		}
@@ -393,22 +408,6 @@ func (q *Query) entityAt(index int) Entity {
 	var count uint32 = 0
 	idx := uint32(index)
 
-	if q.isBatch {
-		batch := q.nodeArchetypes.(*batchArchetypes)
-		nArch := batch.Len()
-		var j int32
-		fmt.Println(batch.StartIndex, batch.EndIndex)
-		for j = 0; j < nArch; j++ {
-			ln := batch.EndIndex[j] - batch.StartIndex[j]
-			if idx < count+ln {
-				fmt.Println(batch.StartIndex[j], batch.EndIndex[j], batch.StartIndex[j]+idx-count)
-				return batch.Archetype[j].GetEntity(batch.StartIndex[j] + idx - count)
-			}
-			count += ln
-		}
-		panic(fmt.Sprintf("query index out of range: index %d, length %d", index, count))
-	}
-
 	if q.isFiltered {
 		ln := int32(len(q.archetypes))
 		var i int32
@@ -416,6 +415,20 @@ func (q *Query) entityAt(index int) Entity {
 			ln := q.archetypes[i].Len()
 			if idx < count+ln {
 				return q.archetypes[i].GetEntity(idx - count)
+			}
+			count += ln
+		}
+		panic(fmt.Sprintf("query index out of range: index %d, length %d", index, count))
+	}
+
+	if q.isBatch {
+		batch := q.nodeArchetypes.(*batchArchetypes)
+		nArch := batch.Len()
+		var j int32
+		for j = 0; j < nArch; j++ {
+			ln := batch.EndIndex[j] - batch.StartIndex[j]
+			if idx < count+ln {
+				return batch.Archetype[j].GetEntity(batch.StartIndex[j] + idx - count)
 			}
 			count += ln
 		}
@@ -430,8 +443,7 @@ func (q *Query) entityAt(index int) Entity {
 		if !nd.HasRelation {
 			// There should be at least one archetype.
 			// Otherwise, the node would be inactive.
-			arches := nd.Archetypes()
-			arch := arches.Get(0)
+			arch := nd.Archetypes().Get(0)
 
 			ln := arch.Len()
 			if idx < count+ln {